Update README.md
Some of the dependencies do not compile on the stable release of the rust toolchain mostly because they are using experimental or non-standardized features that are only available in nightly releases. I faced this issue so I added information to that effect.

If this is not the right spot for this info please place it in the spot or let me know where I should add it.